jarvis/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Voice processing
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Free AI alternatives
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
requests==2.31.0

# Web services
aiohttp==3.9.0
python-dotenv==1.0.0

# Automation
schedule==1.2.0

# Audio processing
sounddevice==0.4.6
numpy==1.24.3

# Web scraping (for news/search)
beautifulsoup4==4.12.2
feedparser==6.0.10
wikipedia==1.4.0

# CORS for web frontend
fastapi-cors==0.0.6